:recycle: wip refactoring deserializer & normalizer
