+++ binding.gyp
@@ -15,8 +15,10 @@
         "src/deserializer.c",
         "src/deserializer_space.c",
         "src/deserializer_opts.c",
+        "src/formater.c",
         "src/normalizer/normalizer_rgb.c",
-        "src/normalizer/regular.c",
+        "src/normalizer/normalizer_regular.c",
+        "src/normalizer/normalizer_space.c",
         "src/normalizer/normalizer_xyz.c",
       ],
       "include_dirs": [
