+++ src/include/deserializer_space.h
@@ -13,6 +13,8 @@
 #include <node_api.h>
 #include "format_props.h"
 
+extern napi_env envglobal;
+
 /**
  * @brief struct that will be used to calculate value
  */
@@ -23,7 +25,6 @@ typedef struct BridgeSpaceObj {
     double clamp;
 } BridgeSpaceObj;
 
-
 /**
  * @brief deserialize space color
  * @param env napi_env
@@ -31,14 +32,5 @@ typedef struct BridgeSpaceObj {
  * @return BridgeSpaceObj pointer
  */
 BridgeSpaceObj *deserializeSpace(napi_env env, napi_value obj);
-
-
-/**
- * @brief normalize the value of regular color for being prepare into being convet into RGB (broken english am sick).
- * @param env napi_env
- * @param obj napi_value
- * @return BridgeObj pointer
- */
-BridgeSpaceObj *normalizeSpace(napi_env env, napi_value obj);
 
 #endif /* deserializer_space_h */
